Update: Network.java wasn't added in previous commit.
Added networking to the code (moved all networking elements to Network class).  For now, it assumes only 2 players are playing.    Right now, each player initializes his own player's location and sends it to other phone.  Server initializes the tokens' locations and sends to clients.  Each players actions are appropriately sent to other device - a small lag, but otherwise ok.

Note: there is a lot of extra commented out code and messy code at this point.  But, I wanted to commit so I wouldn't lose this code.  I'll be fixing this messy code in the next revision as well as generalizing it to allow for n players.